+++ libs/dot-components/src/lib/components/avatar-group/AvatarGroup.spec.tsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import { screen } from '@testing-library/react';
+import { renderWithTheme as render } from '../../testing-utils/RenderWithTheme';
+import { AvatarGroupProps, DotAvatarGroup } from './AvatarGroup';
+import { sampleAvatars } from './AvatarGroup.stories.data';
+
+describe('AvatarGroup', () => {
+  it('should have unchanged API', () => {
+    const props = {
+      avatars: sampleAvatars,
+      max: 3,
+      spacing: 'medium',
+    };
+    const avatarGroupProps: AvatarGroupProps = {
+      avatars: sampleAvatars,
+      max: 3,
+      spacing: 'medium',
+    };
+    expect(avatarGroupProps).toEqual(props);
+  });
+
+  it('should render successfully', () => {
+    const { baseElement } = render(<DotAvatarGroup avatars={sampleAvatars} />);
+    expect(baseElement).toBeTruthy();
+  });
+
+  it('should display the correct number of avatars', () => {
+    render(<DotAvatarGroup avatars={sampleAvatars} max={3} />);
+    const avatars = screen.getAllByTestId('test-avatar');
+
+    expect(avatars).toHaveLength(2);
+  });
+});