+++ libs/dot-components/src/lib/components/avatar-group/AvatarGroup.stories.data.tsx
@@ -1,0 +1,29 @@
+import { AvatarProps } from '../avatar/Avatar';
+
+export const sampleAvatars: Array<AvatarProps> = [
+  {
+    alt: 'avatar alt text',
+    'data-testid': 'test-avatar',
+    imageSrc:
+      'https://cdn1-www.superherohype.com/assets/uploads/2013/11/batmane3-1.jpg',
+    type: 'image',
+  },
+  {
+    alt: 'avatar alt text',
+    'data-testid': 'test-avatar',
+    text: 'Superman',
+    type: 'text',
+  },
+  {
+    alt: 'avatar alt text',
+    'data-testid': 'test-avatar',
+    text: 'Flash',
+    type: 'text',
+  },
+  {
+    alt: 'avatar alt text',
+    'data-testid': 'test-avatar',
+    text: 'Wonder Woman',
+    type: 'text',
+  },
+];